+++ scripts/vagrant.sh
@@ -10,12 +10,13 @@ e "Comment out requiretty in /etc/sudoers"
 sed -i 's/^\(Defaults.*requiretty\)/#\1/' /etc/sudoers
 
 e "Installing Vagrant user and keys"
-useradd vagrant
+# useradd vagrant
 mkdir -m 0700 -p /home/vagrant/.ssh
 cd /home/vagrant/.ssh
 wget --no-check-certificate 'http://github.com/mitchellh/vagrant/raw/master/keys/vagrant.pub' -O authorized_keys
 chmod 600 /home/vagrant/.ssh/authorized_keys
 chown -R vagrant:vagrant /home/vagrant/.ssh
+chmod -R go-rwsx /home/vagrant/.ssh
 
 e "Allow user Vagrant to use sudo without entering a password"
 usermod -G wheel vagrant > /dev/null 2>&1